Add tests and handlers for equality and hashcode based on state
