+++ src/main/java/info/kfgodel/dyna/impl/proxy/handlers/EqualsMethodHandler.java
@@ -1,36 +1,15 @@
 package info.kfgodel.dyna.impl.proxy.handlers;
 
-import info.kfgodel.dyna.impl.proxy.invocation.DynaMethodInvocationHandler;
-import info.kfgodel.dyna.impl.proxy.invocation.DynaTypeMethodInvocation;
-import info.kfgodel.dyna.impl.proxy.invocation.result.HandledResult;
-import info.kfgodel.dyna.impl.proxy.invocation.result.HandlingResult;
-import info.kfgodel.dyna.impl.proxy.invocation.result.UnhandledResult;
-
 /**
  * This class implements the equals method by comparing object states by identity.<br>
  *   Two objects are equal if they have the same state
  *
  * Date: 05/05/19 - 14:38
  */
-public class EqualsMethodHandler implements DynaMethodInvocationHandler {
+public class EqualsMethodHandler extends EqualiltyMethodHandlerSupport  {
 
-  public static final String EQUALS_METHOD_NAME = "equals";
-
-  @Override
-  public HandlingResult tryToHandle(DynaTypeMethodInvocation invocation) {
-    if(EQUALS_METHOD_NAME.equals(invocation.getMethodName()) && invocation.getArgumentCount() == 1){
-      boolean areEquals = calculateEqualityBetween(invocation.getInvokedProxy(), invocation.getArgument(0));
-      return HandledResult.create(areEquals);
-    }
-    return UnhandledResult.instance();
-  }
-
-  private boolean calculateEqualityBetween(Object object, Object other) {
-    return object == other;
-  }
-
-  public static EqualsMethodHandler create() {
-    EqualsMethodHandler handler = new EqualsMethodHandler();
+  public static EqualiltyMethodHandlerSupport create() {
+    EqualiltyMethodHandlerSupport handler = new EqualsMethodHandler();
     return handler;
   }
